+++ docs/changelog.md
@@ -18,6 +18,7 @@ tag: updated
 
 - **Breaking:** Swapped enums for strings for tracking game settings. Prep work for YAML-based game import/export.
 - **Breaking:** `OwnerID` replaces `LocationID` in blocks.
+- New fields for blocks to support reports in the future.
 
 ### Removed